refactor: decouple sdk/wallet-standard from other packages, only depends on @core
