+++ apps/wallet/browser/kernel/utils/middleware.ts
@@ -1,11 +1,12 @@
-import type { ConnectOptions, Networks } from '@walless/core';
+import type { Networks } from '@walless/core';
 import {
 	PopupType,
 	RequestType,
 	ResponseCode,
 	ResponseMessage,
-} from '@walless/messaging';
+} from '@walless/core';
 import { utils } from '@walless/network';
+import type { ConnectOptions } from '@walless/sdk';
 import type { TrustedDomainDocument } from '@walless/store';
 import { selectors } from '@walless/store';
