+++ apps/wallet/src/features/Send/ConfirmTransaction/index.tsx
@@ -2,10 +2,9 @@ import type { FC } from 'react';
 import { useState } from 'react';
 import { ActivityIndicator, StyleSheet } from 'react-native';
 import type { Token } from '@walless/core';
-import { logger } from '@walless/core';
+import { logger, ResponseCode } from '@walless/core';
 import type { SliderHandle } from '@walless/gui';
 import { View } from '@walless/gui';
-import { ResponseCode } from '@walless/messaging';
 import { NavButton } from 'components/NavButton';
 import { showError } from 'modals/Error';
 import { createAndSend, prepareTransactionPayload } from 'utils/transaction';
