+++ apps/wallet/src/features/Widget/BuiltInNetwork/AptosTokensTab/PendingTokens.tsx
@@ -1,9 +1,8 @@
 import type { FC } from 'react';
 import { Image, StyleSheet } from 'react-native';
 import type { AptosPendingToken } from '@walless/core';
-import { shortenAddress } from '@walless/core';
+import { RequestType, shortenAddress } from '@walless/core';
 import { Button, Text, View } from '@walless/gui';
-import { RequestType } from '@walless/messaging';
 import type { aptosHandler } from '@walless/network';
 import { showRequirePasscodeModal } from 'modals/RequirePasscode';
 import { aptosState } from 'state/assets';
