+++ apps/wallet/src/modals/RequirePasscode.tsx
@@ -1,6 +1,7 @@
 import type { FC } from 'react';
 import { useState } from 'react';
 import { Image, StyleSheet } from 'react-native';
+import { ResponseCode } from '@walless/core';
 import type { ModalConfigs } from '@walless/gui';
 import {
 	BindDirections,
@@ -10,7 +11,6 @@ import {
 	View,
 } from '@walless/gui';
 import type { ResponsePayload } from '@walless/messaging';
-import { ResponseCode } from '@walless/messaging';
 
 import { ModalId } from './types';
