+++ packages/core/index.ts
@@ -1,10 +1,10 @@
-export * from './utils/commonTypes';
-export * from './utils/entityTypes';
-export * from './utils/extensionTypes';
+export * from './utils/common';
+export * from './utils/entity';
+export * from './utils/extension';
 export * from './utils/format';
 export * from './utils/helper';
 export * from './utils/logger';
+export * from './utils/messaging';
 export * from './utils/platform';
-export * from './utils/walletTypes';
 export * from './utils/wasm';
-export * from './utils/widgetTypes';
+export * from './utils/widget';
