+++ packages/core/utils/extension.ts
