+++ packages/sdk/package.json
@@ -14,6 +14,7 @@
     "@walless/core": "*"
   },
   "devDependencies": {
+    "@solana/web3.js": "1.78.4",
     "eslint": "^8.56.0"
   }
 }
