+++ packages/sdk/providers/common.ts
@@ -1,8 +1,7 @@
-import type { ConnectOptions } from '@walless/core';
-import { Timeout } from '@walless/core';
-import { RequestType } from '@walless/messaging';
+import { RequestType, Timeout } from '@walless/core';
 
 import { sendRequest } from '../utils/messaging';
+import type { ConnectOptions } from '../utils/type';
 
 const from = 'walless@sdk';
